backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
python-dotenv==1.0.1
